Upload page restructured with notice display and upload fixed
