+++ client/src/components/auth/AuthOptions.js
@@ -1,6 +1,6 @@
 import React, { useContext } from "react";
 import { useHistory } from "react-router-dom";
-import UserContext from "../../context/UserContext";
+import UserContext from "../../global/UserContext";
 
 export default function AuthOptions() {
   const { userData, setUserData } = useContext(UserContext);
