+++ client/src/components/pages/DisplayNotice/style.module.css
@@ -1,0 +1,16 @@
+.noticeSectionContainer {
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  justify-content: space-evenly;
+}
+
+.noNoticeWarning {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+.noNoticeWarning h1 {
+  margin: 0;
+}
